some touchups
